Cleanup Xcode warnings

+++ MediaFireSDK/MediaFireSDK/requestmanagers/MFSerialRequestManager.m
@@ -514,7 +514,7 @@ static id instance = nil;
     
     NSMutableArray* orderedParams = [MFSerialRequestManager getOrderedArrayOfParameters:newParams];
     
-    signature = [NSString stringWithFormat:@"%lu%@%@?%@", (secret % 256), tokenPacket[@"time"], baseUrl, [MFSerialRequestManager getQueryStringFromOrderedParameters:orderedParams]];
+    signature = [NSString stringWithFormat:@"%lu%@%@?%@", (unsigned long)(secret % 256), tokenPacket[@"time"], baseUrl, [MFSerialRequestManager getQueryStringFromOrderedParameters:orderedParams]];
     
     NSString * sig = [MFHash md5Hex:signature];
     [MFSerialRequestManager urlEncodeOrderedParameters:orderedParams];
